# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/courses": {
            "get": {
                "description": "Get all courses",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "List all courses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.Course"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/courses/{id}": {
            "get": {
                "description": "Get a course by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "Get a course by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Course"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a course by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "Delete a course by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Course"
                        }
                    }
                }
            }
        },
        "/api/v1/lectures": {
            "get": {
                "description": "Get all lectures",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lectures"
                ],
                "summary": "List all lectures",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.Teacher"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a lecture",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lectures"
                ],
                "summary": "Create a lecture",
                "parameters": [
                    {
                        "description": "Lecture object",
                        "name": "lecture",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Teacher"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.Teacher"
                        }
                    }
                }
            }
        },
        "/api/v1/lectures/{id}": {
            "get": {
                "description": "Get a lecture by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lectures"
                ],
                "summary": "Get a lecture by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lecture ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Teacher"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a lecture by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lectures"
                ],
                "summary": "Update a lecture by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lecture ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Lecture object",
                        "name": "lecture",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Teacher"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Teacher"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a lecture by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lectures"
                ],
                "summary": "Delete a lecture by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lecture ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/v1/students": {
            "get": {
                "description": "Get all students",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "List all students",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.Student"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "Create a student",
                "parameters": [
                    {
                        "description": "Student object",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Student"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.Student"
                        }
                    }
                }
            }
        },
        "/api/v1/students/{id}": {
            "get": {
                "description": "Get a student by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "Get a student by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Student"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a student by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "Update a student by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Student object",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Student"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.Student"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a student by ID",
                "tags": [
                    "Students"
                ],
                "summary": "Delete a student by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.Course": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "lectures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.Teacher"
                    }
                },
                "name": {
                    "type": "string"
                },
                "semester": {
                    "type": "string"
                },
                "students": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.Student"
                    }
                }
            }
        },
        "controllers.Student": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "controllers.Teacher": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}